Implemented performance test for 'get_line_data' & 'get_kpi_data_viz'
